Revert "Add poseidon2 compressed"
This reverts commit e11fc2123364d32dabe3850c70487c6680a24771.
