+++ poseidon2_plonky2/src/lib.rs
@@ -1,13 +1,7 @@
 
 extern crate alloc;
 
-#[macro_use]
-mod macros;
 pub mod poseidon2_hash;
-pub mod poseidon2_compressed_hash;
 pub mod poseidon2_gate;
-pub mod poseidon2_compressed_gate;
 pub mod poseidon2_goldilock;
-pub mod poseidon2_compressed_goldilock;
-
