+++ poseidon2_plonky2/src/poseidon2_hash.rs
@@ -130,13 +130,389 @@ pub const ALL_ROUND_CONSTANTS: [u64; MAX_WIDTH * N_ROUNDS]  = [
     0xc0c2e289afa75181, 0x1e4fa2ad948978b7, 0x2a226a127a0bb26a, 0xe61738a70357ce76,
 ];
 
-implement_poseidon2!(Poseidon2, ALL_ROUND_CONSTANTS, SPONGE_WIDTH);
+const WIDTH: usize = SPONGE_WIDTH;
+pub trait Poseidon2: PrimeField64 {
+    // Total number of round constants required: width of the input
+    // times number of rounds.
+    const N_ROUND_CONSTANTS: usize = WIDTH * N_ROUNDS;
+
+    // We only need INTERNAL_MATRIX_DIAG_M_1 here, specifying the diagonal - 1 of the internal matrix
+    const INTERNAL_MATRIX_DIAG_M_1: [u64; WIDTH];
+
+    #[inline(always)]
+    #[unroll_for_loops]
+    fn external_matrix(state_: &mut[Self; WIDTH]) {
+        // Applying cheap 4x4 MDS matrix to each 4-element part of the state
+        // The matrix in this case is:
+        // M_4 =
+        // [5   7   1   3]
+        // [4   6   1   1]
+        // [1   3   5   7]
+        // [1   1   4   6]
+        // The computation is shown in more detail in https://tosc.iacr.org/index.php/ToSC/article/view/888/839, Figure 13 (M_{4,4}^{8,4} with alpha = 2)
+        let t4 = WIDTH / 4;
+        let mut state_u128: Vec<u128> = state_.iter().map(|&e| e.to_canonical_u64() as u128).collect();
+        for i in 0..t4 {
+            let start_index = i * 4;
+            let mut t_0 = state_u128[start_index];
+            t_0 += state_u128[start_index + 1];
+            let mut t_1 = state_u128[start_index + 2];
+            t_1 += state_u128[start_index + 3];
+            let mut t_2 = state_u128[start_index + 1];
+            t_2 = t_2 + t_2;
+            t_2 += t_1;
+            let mut t_3 = state_u128[start_index + 3];
+            t_3 = t_3 + t_3;
+            t_3 += t_0;
+            let mut t_4 = t_1;
+            t_4 = 4 * t_4;
+            t_4 += t_3;
+            let mut t_5 = t_0;
+            t_5 = 4 * t_5;
+            t_5 += t_2;
+            let mut t_6 = t_3;
+            t_6 += t_5;
+            let mut t_7 = t_2;
+            t_7 += t_4;
+            state_u128[start_index] = t_6;
+            state_u128[start_index + 1] = t_5;
+            state_u128[start_index + 2] = t_7;
+            state_u128[start_index + 3] = t_4;
+        }
+
+        // Applying second cheap matrix
+        // This completes the multiplication by
+        // M_E =
+        // [2*M_4    M_4    M_4]
+        // [  M_4  2*M_4    M_4]
+        // [  M_4    M_4  2*M_4]
+        // using the results with M_4 obtained above
+        let mut stored = [0 as u128; 4];
+        for l in 0..4 {
+            stored[l] = state_u128[l];
+            for j in 1..t4 {
+                stored[l] += state_u128[4 * j + l];
+            }
+        }
+        for i in 0..state_u128.len() {
+            state_u128[i] += stored[i % 4];
+            state_[i] = Self::from_noncanonical_u128(state_u128[i]);
+        }
+    }
+
+    /// Same as `external_matrix` for field extensions of `Self`.
+    fn external_matrix_field<F: FieldExtension<D, BaseField = Self>, const D: usize>(
+        state: &mut [F; WIDTH],
+    ) {
+        // Applying cheap 4x4 MDS matrix to each 4-element part of the state
+        let t4 = WIDTH / 4;
+        for i in 0..t4 {
+            let start_index = i * 4;
+            let mut t_0 = state[start_index];
+            t_0 += state[start_index + 1];
+            let mut t_1 = state[start_index + 2];
+            t_1 += state[start_index + 3];
+            let mut t_2 = state[start_index + 1];
+            t_2 = t_2 + t_2;
+            t_2 += t_1;
+            let mut t_3 = state[start_index + 3];
+            t_3 = t_3 + t_3;
+            t_3 += t_0;
+            let mut t_4 = t_1;
+            t_4 = F::from_canonical_u64(4) * t_4;
+            t_4 += t_3;
+            let mut t_5 = t_0;
+            t_5 = F::from_canonical_u64(4) * t_5;
+            t_5 += t_2;
+            let mut t_6 = t_3;
+            t_6 += t_5;
+            let mut t_7 = t_2;
+            t_7 += t_4;
+            state[start_index] = t_6;
+            state[start_index + 1] = t_5;
+            state[start_index + 2] = t_7;
+            state[start_index + 3] = t_4;
+        }
+
+        // Applying second cheap matrix
+        let mut stored = [F::ZERO; 4];
+        for l in 0..4 {
+            stored[l] = state[l];
+            for j in 1..t4 {
+                stored[l] += state[4 * j + l];
+            }
+        }
+        for i in 0..state.len() {
+            state[i] += stored[i % 4];
+        }
+
+    }
+
+    /// Recursive version of `external_matrix`.
+    fn external_matrix_circuit<const D: usize>(
+        builder: &mut CircuitBuilder<Self, D>,
+        state: &mut [ExtensionTarget<D>; WIDTH],
+    )
+        where
+            Self: RichField + Extendable<D>,
+    {
+        // In contrast to the Poseidon circuit, we *may not need* PoseidonMdsGate, because the number of constraints will fit regardless
+        // Check!
+        let four = Self::from_canonical_u64(0x4);
+        // let four = builder.constant_extension(Self::Extension::from_canonical_u64(0x4));
+
+        // Applying cheap 4x4 MDS matrix to each 4-element part of the state
+        let t4 = WIDTH / 4;
+        for i in 0..t4 {
+            let start_index = i * 4;
+            let mut t_0 = state[start_index];
+            t_0 = builder.add_extension(t_0, state[start_index + 1]);
+            let mut t_1 = state[start_index + 2];
+            t_1 = builder.add_extension(t_1, state[start_index + 3]);
+            let mut t_2 = state[start_index + 1];
+            t_2 = builder.add_extension(t_2, t_2); // Double
+            t_2 = builder.add_extension(t_2, t_1);
+            let mut t_3 = state[start_index + 3];
+            t_3 = builder.add_extension(t_3, t_3); // Double
+            t_3 = builder.add_extension(t_3, t_0);
+            let mut t_4 = t_1;
+            t_4 = builder.mul_const_extension(four, t_4); // times 4
+            t_4 = builder.add_extension(t_4, t_3);
+            let mut t_5 = t_0;
+            t_5 = builder.mul_const_extension(four, t_5); // times 4
+            t_5 = builder.add_extension(t_5, t_2);
+            let mut t_6 = t_3;
+            t_6 = builder.add_extension(t_6, t_5);
+            let mut t_7 = t_2;
+            t_7 = builder.add_extension(t_7, t_4);
+            state[start_index] = t_6;
+            state[start_index + 1] = t_5;
+            state[start_index + 2] = t_7;
+            state[start_index + 3] = t_4;
+        }
+
+        // Applying second cheap matrix
+        let mut stored = [builder.zero_extension(); 4];
+        for l in 0..4 {
+            stored[l] = state[l];
+            for j in 1..t4 {
+                stored[l] = builder.add_extension(stored[l], state[4 * j + l]);
+            }
+        }
+        for i in 0..state.len() {
+            state[i] = builder.add_extension(state[i], stored[i % 4]);
+        }
+    }
+
+    #[inline(always)]
+    #[unroll_for_loops]
+    fn internal_matrix(state_: &mut [Self; WIDTH]) {
+
+        // This computes the mutliplication with the matrix
+        // M_I =
+        // [r_1     1   1   ...     1]
+        // [  1   r_2   1   ...     1]
+        // ...
+        // [  1     1   1   ...   r_t]
+        // for pseudo-random values r_1, r_2, ..., r_t. Note that for efficiency in Self::INTERNAL_MATRIX_DIAG_M_1 only r_1 - 1, r_2 - 1, ..., r_t - 1 are stored
+        // Compute input sum
+        let mut sum = state_[0].to_noncanonical_u64() as u128;
+        state_
+            .iter()
+            .skip(1)
+            .take(WIDTH-1)
+            .for_each(|el| sum += (*el).to_noncanonical_u64() as u128);
+        let f_sum = Self::from_noncanonical_u128(sum);
+        // Add sum + diag entry * element to each element
+        for i in 0..WIDTH {
+            state_[i] *= Self::from_canonical_u64(Self::INTERNAL_MATRIX_DIAG_M_1[i]);
+            state_[i] += f_sum;
+        }
+    }
+
+    /// Same as `internal_matrix` for field extensions of `Self`.
+    fn internal_matrix_field<F: FieldExtension<D, BaseField = Self>, const D: usize>(
+        state: &mut [F; WIDTH],
+    ) {
+        let t = WIDTH;
+
+        // Compute input sum
+        let mut sum = state[0];
+        state
+            .iter()
+            .skip(1)
+            .take(t-1)
+            .for_each(|el| sum += *el);
+        // Add sum + diag entry * element to each element
+        for i in 0..state.len() {
+            state[i] *= F::from_canonical_u64(Self::INTERNAL_MATRIX_DIAG_M_1[i]);
+            state[i] += sum;
+        }
+    }
+
+    /// Recursive version of `internal_matrix`.
+    fn internal_matrix_circuit<const D: usize>(
+        builder: &mut CircuitBuilder<Self, D>,
+        state: &mut [ExtensionTarget<D>; WIDTH],
+    )
+        where
+            Self: RichField + Extendable<D>,
+    {
+        // In contrast to the Poseidon circuit, we *may not need* PoseidonMdsGate, because the number of constraints will fit regardless
+        // Check!
+
+        // Compute input sum
+        let mut sum = state[0];
+        for i in 1..state.len() {
+            sum = builder.add_extension(sum, state[i]);
+        }
+        // Add sum + diag entry * element to each element
+        for i in 0..state.len() {
+            // Computes `C * x + y`
+            state[i] = builder.mul_const_add_extension(Self::from_canonical_u64(<Self as Poseidon2>::INTERNAL_MATRIX_DIAG_M_1[i]), state[i], sum);
+        }
+    }
+
+    #[inline(always)]
+    #[unroll_for_loops]
+    fn constant_layer(state: &mut [Self; WIDTH], round_ctr: usize) {
+        for i in 0..12 {
+            if i < WIDTH {
+                let round_constant = ALL_ROUND_CONSTANTS[i + WIDTH * round_ctr];
+                unsafe {
+                    state[i] = state[i].add_canonical_u64(round_constant);
+                }
+            }
+        }
+    }
+
+    /// Same as `constant_layer` for field extensions of `Self`.
+    fn constant_layer_field<F: FieldExtension<D, BaseField = Self>, const D: usize>(
+        state: &mut [F; WIDTH],
+        round_ctr: usize,
+    ) {
+        for i in 0..WIDTH {
+            state[i] += F::from_canonical_u64(ALL_ROUND_CONSTANTS[i + WIDTH * round_ctr]);
+        }
+    }
+
+    /// Recursive version of `constant_layer`.
+    fn constant_layer_circuit<const D: usize>(
+        builder: &mut CircuitBuilder<Self, D>,
+        state: &mut [ExtensionTarget<D>; WIDTH],
+        round_ctr: usize,
+    ) where
+        Self: RichField + Extendable<D>,
+    {
+        for i in 0..WIDTH {
+            let c = ALL_ROUND_CONSTANTS[i + WIDTH * round_ctr];
+            let c = Self::Extension::from_canonical_u64(c);
+            let c = builder.constant_extension(c);
+            state[i] = builder.add_extension(state[i], c);
+        }
+    }
+
+    #[inline(always)]
+    fn sbox_monomial<F: FieldExtension<D, BaseField = Self>, const D: usize>(x: F) -> F {
+        // x |--> x^7
+        let x2 = x.square();
+        let x4 = x2.square();
+        let x3 = x * x2;
+        x3 * x4
+    }
+
+    /// Recursive version of `sbox_monomial`.
+    fn sbox_monomial_circuit<const D: usize>(
+        builder: &mut CircuitBuilder<Self, D>,
+        x: ExtensionTarget<D>,
+    ) -> ExtensionTarget<D>
+        where
+            Self: RichField + Extendable<D>,
+    {
+        // x |--> x^7
+        builder.exp_u64_extension(x, 7)
+    }
+
+    #[inline(always)]
+    #[unroll_for_loops]
+    fn sbox_layer(state: &mut [Self; WIDTH]) {
+        for i in 0..12 {
+            if i < WIDTH {
+                state[i] = Self::sbox_monomial(state[i]);
+            }
+        }
+    }
+
+    /// Same as `sbox_layer` for field extensions of `Self`.
+    fn sbox_layer_field<F: FieldExtension<D, BaseField = Self>, const D: usize>(
+        state: &mut [F; WIDTH],
+    ) {
+        for i in 0..WIDTH {
+            state[i] = Self::sbox_monomial(state[i]);
+        }
+    }
+
+    /// Recursive version of `sbox_layer`.
+    fn sbox_layer_circuit<const D: usize>(
+        builder: &mut CircuitBuilder<Self, D>,
+        state: &mut [ExtensionTarget<D>; WIDTH],
+    ) where
+        Self: RichField + Extendable<D>,
+    {
+        for i in 0..WIDTH {
+            state[i] = <Self as Poseidon2>::sbox_monomial_circuit(builder, state[i]);
+        }
+    }
+
+    #[inline]
+    fn full_rounds(state: &mut [Self; WIDTH], round_ctr: &mut usize) {
+        for _ in 0..HALF_N_FULL_ROUNDS {
+            Self::constant_layer(state, *round_ctr);
+            Self::sbox_layer(state);
+            Self::external_matrix(state);
+            *round_ctr += 1;
+        }
+    }
+
+    #[inline]
+    fn partial_rounds(state: &mut [Self; WIDTH], round_ctr: &mut usize) {
+        let mut constant_counter = HALF_N_FULL_ROUNDS * WIDTH;
+        for _ in 0..N_PARTIAL_ROUNDS {
+            unsafe {
+                state[0] = state[0].add_canonical_u64(ALL_ROUND_CONSTANTS[constant_counter]);
+                constant_counter += WIDTH;
+            }
+            state[0] = Self::sbox_monomial(state[0]);
+            Self::internal_matrix(state);
+        }
+        *round_ctr += N_PARTIAL_ROUNDS;
+    }
+
+    #[inline]
+    fn poseidon2(input: [Self; WIDTH]) -> [Self; WIDTH] {
+        let mut state = input;
+        let mut round_ctr = 0;
+
+        // First external matrix
+        Self::external_matrix(&mut state);
+
+        Self::full_rounds(&mut state, &mut round_ctr);
+        Self::partial_rounds(&mut state, &mut round_ctr);
+        Self::full_rounds(&mut state, &mut round_ctr);
+        debug_assert_eq!(round_ctr, N_ROUNDS);
+
+        state
+    }
+}
 
 pub struct Poseidon2Permutation;
 impl<F: RichField + Poseidon2> PlonkyPermutation<F> for Poseidon2Permutation {
     fn permute(input: [F; SPONGE_WIDTH]) -> [F; SPONGE_WIDTH] {
         F::poseidon2(input)
     }
+    /*fn permute_c(_: [F; COMPRESSION_WIDTH]) -> [F; COMPRESSION_WIDTH] {
+        [F::ZERO; COMPRESSION_WIDTH]
+    }*/
 }
 
 /// Poseidon2 hash function.
